requirements.txt
-r app/requirements.txt
requests
pyyaml
dockerfile_parse